requirements.txt
asgiref==3.3.4
Django==3.2.4
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.4
drf-yasg==1.20.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.1.0
swagger-spec-validator==2.7.3
Celery==5.1.0
redis==3.5.3
jsonschema~=3.2.0
PyJWT==1.7.1
flower==0.9.7
coverage==5.5
model-bakery==1.3.1